+++ test/integration/test_cases/web_content/cookbook/interactive/mouse_wheel_zoom.mjs
@@ -53,7 +53,7 @@ const testSteps = [
 
 		let throttle = new Throttle();
 
-		chart._container.addEventListener('wheel', event => {
+		chart.getCanvasElement().addEventListener('wheel', event => {
 			event.preventDefault();
 		})