+++ test/integration/test_cases/web_content/cookbook/interactive/seeker_slider.mjs
@@ -18,7 +18,7 @@ const testSteps = [
 		input.min=0;
 		input.max=1000;
 		input.style="width: 500px; position: absolute; top: 540px; left: 30px;";
-		chart._container.parentElement.appendChild(input);
+		chart.getCanvasElement().parentElement.appendChild(input);
 
 		let animation = chart.animate({
 			x: 'Country',