+++ CHANGELOG.md
@@ -13,6 +13,7 @@
   triggered on.
 - Drawing events are containing drawing geometry in an improved format.
 - Missing TypeScript definitions for Event objects added.
+- getCanvasElement() introduced to expose the underlying HTML canvas element.
 
 ## [0.8.1] - 2023-08-15